refactor(gradle-inspector): Migrate the code to use the dependency graph
See [1] for context.

[1]: https://github.com/oss-review-toolkit/ort/issues/3825
