+++ plugins/package-managers/gradle-inspector/src/funTest/assets/projects/synthetic/gradle-expected-output-lib-without-repo.yml
@@ -15,15 +15,7 @@ project:
     revision: "<REPLACE_REVISION>"
     path: "plugins/package-managers/gradle/src/funTest/assets/projects/synthetic/gradle/lib-without-repo"
   homepage_url: ""
-  scopes:
-  - name: "compileClasspath"
-    dependencies: []
-  - name: "runtimeClasspath"
-    dependencies: []
-  - name: "testCompileClasspath"
-    dependencies: []
-  - name: "testRuntimeClasspath"
-    dependencies: []
+  scopes: []
 packages: []
 issues:
 - timestamp: "1970-01-01T00:00:00Z"
